requirements.txt
asgiref==3.4.1
Django==3.2.9
django-cors-headers==3.10.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
install==1.3.4
psycopg==3.0.3
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
psycopg2